data_model/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pandas